+++ DebugLocalizationDemo/UIKitDemoView.swift
@@ -2,10 +2,8 @@ import SwiftUI
 import DebugLocalizationCore
 
 struct UIKitDemoView: UIViewControllerRepresentable {
-    let localizer: DebugLocalizer
-
     func makeUIViewController(context: Context) -> UINavigationController {
-        let controller = UIKitDemoViewController(localizer: localizer)
+        let controller = UIKitDemoViewController()
         return UINavigationController(rootViewController: controller)
     }