+++ frameworks/DebugLocalizationPackage/Sources/DebugLocalizationCore/DebugLocalizationProvider.swift
@@ -1,7 +1,17 @@
 import Foundation
 
-public protocol DebugLocalizationProvider: Sendable {
-    func localize(_ text: String, into languageIdentifier: String) async throws -> String
+public protocol LocalizationProvider: Sendable {
+    func translate(_ text: String) async -> String
+}
+
+public protocol SyncLocalizationProvider: LocalizationProvider {
+    func translateSync(_ text: String) -> String
+}
+
+public extension SyncLocalizationProvider {
+    func translate(_ text: String) async -> String {
+        translateSync(text)
+    }
 }
 
 public enum DebugLocalizationError: Error {